"Added tooltip title and data-toggle attribute to 'Add Result' button in class\_wise\_students\_list.html; Added variable and if condition in sidebar\_template.html for staff announcements"
This commit message describes the changes made in the provided code diff. The first part mentions the addition of a tooltip title and data-toggle attribute to the 'Add Result' button in class\_wise\_students\_list.html. The second part mentions the addition of a variable and if condition in sidebar\_template.html for staff announcements. The message is concise and sticks to describing the changes without making any interpretations.